# design/05-api/OpenAPI.yaml
openapi: 3.0.3
info:
  title: 和悦医养服务平台 API
  version: 0.2.0
  description: 第一阶段 OA 基座、IAM、物资管理 REST API。
servers:
  - url: /api/v1
paths:
...
  /oa/instances/claimable-materials/search:
    get:
      operationId: searchClaimableMaterials
      summary: 物品领用可选物资
      description: 按 warehouseId 对应仓库类型过滤可申领物资；AI 不允许调用。
      tags: [OA]
      parameters:
        - name: warehouseId
          in: query
          required: true
          schema: { type: integer, format: int64 }
        - name: keyword
          in: query
          required: false
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /oa/tasks/{taskId}:
    get:
      operationId: getTaskDetail
      summary: 待办详情
      description: 当前可处理人查看 OA 申请内容；AI 不允许调用。
      tags: [OA]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200': { description: OK }
  /oa/tasks/{taskId}/approve:
    post:
      operationId: approveTask
      summary: 审批通过
      tags: [OA]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200': { description: OK }
  /oa/tasks/{taskId}/reject:
    post:
      operationId: rejectTask
      summary: 审批驳回
      tags: [OA]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200': { description: OK }
  /oa/instances/{instanceId}/urge:
    post:
      operationId: urgeInstance
      summary: 催办流程
      tags: [OA]
      parameters:
        - $ref: '#/components/parameters/InstanceId'
      responses:
        '200': { description: OK }
  /oa/instances/{instanceId}/revoke:
    post:
      operationId: revokeInstance
      summary: 撤销本人发起流程
      description: 仅发起人可撤销运行中或待配置流程；撤销后待办取消，AI 不允许调用。
      tags: [OA]
      parameters:
        - $ref: '#/components/parameters/InstanceId'
      responses:
        '200': { description: OK }
...
  /inventory/items/{itemId}:
    get:
      operationId: getInventoryItem
      summary: 物资档案详情
      description: 需要 inventory:item:read；AI 可只读调用。
      tags: [Inventory]
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { description: OK }
    put:
      operationId: updateInventoryItem
      summary: 修改物资档案
      description: 需要 inventory:item:write；路径 itemId 为唯一修改目标，ID 不允许通过请求体修改；AI 不允许调用。
      tags: [Inventory]
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { description: OK }
    delete:
      operationId: deleteInventoryItem
      summary: 删除物资档案
      description: 需要 inventory:item:write；有库存余额或库存流水/出入库历史引用时禁止删除；AI 不允许调用。
      tags: [Inventory]
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { description: OK }
...
  /inventory/stocks:
    get:
      operationId: listInventoryStocks
      summary: 库存查询
      description: 支持 keyword 参数按库存 ID、仓库编码/名称、物资编码/名称等模糊查询；AI 可只读调用。
      tags: [Inventory]
      parameters:
        - name: keyword
          in: query
          required: false
          schema: { type: string }
          description: 模糊查询关键字
      responses:
        '200': { description: OK }
    post:
      operationId: createInventoryStock
      summary: 新增库存
      description: 需要 inventory:stock:write；AI 不允许调用。库存 ID 由数据库或运行时序列自增生成。
      tags: [Inventory]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [warehouseId, itemId, quantity]
              properties:
                warehouseId: { type: integer, format: int64 }
                itemId: { type: integer, format: int64 }
                quantity: { type: number }
      responses:
        '200': { description: OK }
...
components:
  parameters:
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema: { type: integer, format: int64 }
    InstanceId:
      name: instanceId
      in: path
      required: true
      schema: { type: integer, format: int64 }